Memory Leak & Preference Fixes
Memory leak in Student patched. Changed some other minor parts to fit comparison preferences.

+++ src/include/Rectangle.h
@@ -8,7 +8,7 @@
 
 DECLARE_STRUCT(Rectangle,
 {   GenericTraits *impl;
-    int height;
-    int width; })
+    size_t height;
+    size_t width; })
 
 #endif
